Replace UserStoreConnector with UserStore
