+++ modules/broker-auth/src/main/java/io/ballerina/messaging/broker/auth/authentication/jaas/UserStoreLoginModule.java
@@ -22,7 +22,7 @@ import com.sun.security.auth.UserPrincipal;
 import io.ballerina.messaging.broker.auth.AuthException;
 import io.ballerina.messaging.broker.auth.BrokerAuthConstants;
 import io.ballerina.messaging.broker.auth.authentication.AuthResult;
-import io.ballerina.messaging.broker.auth.user.UserStoreConnector;
+import io.ballerina.messaging.broker.auth.authorization.UserStore;
 
 import java.io.IOException;
 import java.util.Map;
@@ -51,14 +51,14 @@ public class UserStoreLoginModule implements LoginModule {
     private UserPrincipal userPrincipal;
     private Subject subject;
     private CallbackHandler callbackHandler;
-    private UserStoreConnector userStoreConnector;
+    private UserStore userStore;
 
     @Override
     public void initialize(Subject subject, CallbackHandler callbackHandler, Map<String, ?> sharedState,
                            Map<String, ?> options) {
         this.subject = subject;
         this.callbackHandler = callbackHandler;
-        this.userStoreConnector = (UserStoreConnector) options.get(BrokerAuthConstants.PROPERTY_USER_STORE_CONNECTOR);
+        this.userStore = (UserStore) options.get(BrokerAuthConstants.PROPERTY_USER_STORE_CONNECTOR);
     }
 
     @Override
@@ -75,7 +75,7 @@ public class UserStoreLoginModule implements LoginModule {
         }
         userName = userNameCallback.getName();
         password = passwordCallback.getPassword();
-        AuthResult authResult = userStoreConnector.authenticate(userName, password);
+        AuthResult authResult = userStore.authenticate(userName, password);
         success = authResult.isAuthenticated();
         if (success) {
             authenticationId = authResult.getUserId();
